Refactor: 'Kartlegging' funksjoner -> KartleggingTransactional
